use BigInt()
closes https://github.com/porter-finance/frontend/issues/375

https://www.loom.com/share/bc764e4ee5b7493a95c58f4dc667a61b

+++ src/pages/Products/index.tsx
@@ -109,7 +109,7 @@ export const createTable = (data: Bond[]) =>
         </span>
       ),
       cost: clearingPrice
-        ? `${Number(formatUnits(clearingPrice * maxSupply, decimals)).toLocaleString()} ${
+        ? `${Number(formatUnits(BigInt(clearingPrice * maxSupply), decimals)).toLocaleString()} ${
             paymentToken.symbol
           }`
         : '-',
